+++ cmd/neofs-cli/modules/object/range.go
@@ -69,9 +69,9 @@ func getObjectRange(cmd *cobra.Command, _ []string) error {
 	if filename == "" {
 		out = os.Stdout
 	} else {
-		f, err := os.OpenFile(filename, os.O_CREATE|os.O_WRONLY, os.ModePerm)
+		f, err := openFileForPayload(filename)
 		if err != nil {
-			return fmt.Errorf("can't open file '%s': %w", filename, err)
+			return err
 		}
 
 		defer f.Close()